ajout tableau

+++ view/back_office/src/component/UserTable/UserTable.jsx
@@ -11,22 +11,26 @@ export const UserTable = () => {
   }, []);
 
   const handleDelete = (id) => {
-    fetch(`http://localhost/yonnha_site_reservation/api/users/${id}`, {
-      method: 'DELETE',
-      headers: {"Access-Control-Allow-Origin" : "*",
-      "Access-Control-Allow-Methods": "*",
-      "Access-Control-Allow-Headers": "*"}
-    })
-      .then(response => {
-        if (response.ok) {
-          setUsers(users.filter(user => user.id !== id));
-        } else {
-          console.error(`Une erreur est survenue : ${response.status}`);
-        }
+    const confirmed = window.confirm("Êtes-vous sûr de vouloir supprimer cet utilisateur ?");
+  
+    if (confirmed) {
+      fetch(`http://localhost/yonnha_site_reservation/api/users/${id}`, {
+        method: 'DELETE',
+        headers: {"Access-Control-Allow-Origin" : "*",
+          "Access-Control-Allow-Methods": "*",
+          "Access-Control-Allow-Headers": "*"}
       })
-      .catch(error => console.error(error));
+        .then(response => {
+          if (response.ok) {
+            setUsers(users.filter(user => user.id !== id));
+          } else {
+            console.error(`Une erreur est survenue : ${response.status}`);
+          }
+        })
+        .catch(error => console.error(error));
+    }
   }
-
+  
   return (
     <table>
       <thead>
@@ -55,4 +59,4 @@ export const UserTable = () => {
       </tbody>
     </table>
   );
-}
+}
